deploy/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
razorpay==1.4.1
pydantic==2.7.4
